// src/components/Editor/Materials/types.ts
export interface ITextConfigType {
  key: string;
  name: string;
  type: 'Text';
}

export type TTextDefaultType = string;

////////////////////////
export interface ITextAreaConfigType {
  key: string;
  name: string;
  type: 'TextArea';
}

export type TTextAreaDefaultType = string;

////////////////////////////
export interface INumberConfigType {
  key: string;
  name: string;
  type: 'Number';
  range?: [number, number];
  step?: number;
}
// ...
export interface IColorConfigType {
  key: string;
  name: string;
  type: 'Color';
}
// ...
export interface ISwitchConfigType {
  key: string;
  name: string;
  notAntd: string;
  type: 'Switch';
}
export type TSwitchDefaultType = boolean;

/////////////////////////////
export interface ICardPickerConfigType<T> {
  key: string;
  name: string;
  type: 'CardPicker';
  notAntd: string;
  range: {
    icons: Array<T>;
  };
}